requirements.txt
# Selenium scripts (selenium_file/)
selenium>=4.0.0
openpyxl>=3.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
